+++ packages/vseed/src/pipeline/advanced/chart/pipes/reshape/reshapeWithDualEncoding.ts
@@ -17,18 +17,10 @@ export const reshapeWithDualEncoding: AdvancedPipe = (advancedVSeed, context) =>
   const result = { ...advancedVSeed }
   const { vseed } = context
   const { dataset } = vseed as ColumnParallel
-  const { dimensions, measures, encoding, chartType } = advancedVSeed
-
-  if (!measures || !dimensions || !dataset || !encoding) {
-    return result
-  }
-
-  if (measures.length === 0) {
-    throw new Error('measures can not be empty')
-  }
+  const { dimensions = [], measures = [], encoding, chartType } = advancedVSeed
 
   if (measures.length > 2) {
-    throw new Error('measures can not be more than 2')
+    throw new Error('measures can not be more than 2 groups in dualAxis')
   }
   const foldInfoList: FoldInfo[] = []
   const unfoldInfoList: UnfoldInfo[] = []
@@ -36,7 +28,6 @@ export const reshapeWithDualEncoding: AdvancedPipe = (advancedVSeed, context) =>
   const datasets: Dataset[] = []
   const primaryMeasures = measures[0] as MeasureGroup
   const secondaryMeasures = (measures[1] || []) as MeasureGroup
-  const hasEncoding = (vseed.dimensions || []).some((item: Dimension) => item.encoding)
 
   if (primaryMeasures && primaryMeasures.children) {
     const {
@@ -49,7 +40,7 @@ export const reshapeWithDualEncoding: AdvancedPipe = (advancedVSeed, context) =>
       uniqueBy(primaryMeasures.children, (item) => item.id),
       encoding as Encoding,
       {
-        colorItemAsId: hasEncoding,
+        colorItemAsId: false,
         foldMeasureValue: FoldPrimaryMeasureValue,
         colorMeasureId: getColorMeasureId(advancedVSeed as AdvancedVSeed),
       },
@@ -71,6 +62,7 @@ export const reshapeWithDualEncoding: AdvancedPipe = (advancedVSeed, context) =>
       uniqueBy(secondaryMeasures.children, (item) => item.id),
       encoding as Encoding,
       {
+        colorItemAsId: false,
         foldMeasureValue: FoldSecondaryMeasureValue,
         colorMeasureId: getColorMeasureId(advancedVSeed as AdvancedVSeed),
       },