+++ packages/vseed/src/pipeline/advanced/chart/pipes/reshape/reshapeWithScatterEncoding.ts
@@ -1,5 +1,11 @@
 import { uniqueBy } from 'remeda'
-import { dataReshapeByEncoding, FoldXMeasureId, FoldXMeasureValue, FoldYMeasureId, FoldYMeasureValue } from 'src/dataReshape'
+import {
+  dataReshapeByEncoding,
+  FoldXMeasureId,
+  FoldXMeasureValue,
+  FoldYMeasureId,
+  FoldYMeasureValue,
+} from 'src/dataReshape'
 import { getColorMeasureId } from 'src/pipeline/spec/chart/pipes'
 import type {
   AdvancedPipe,
@@ -16,18 +22,10 @@ export const reshapeWithScatterEncoding: AdvancedPipe = (advancedVSeed, context)
   const result = { ...advancedVSeed }
   const { vseed } = context
   const { dataset } = vseed as ColumnParallel
-  const { dimensions, measures, encoding, chartType } = advancedVSeed
-
-  if (!measures || !dimensions || !dataset || !encoding) {
-    return result
-  }
-
-  if (measures.length === 0) {
-    throw new Error('measures can not be empty')
-  }
+  const { dimensions = [], measures = [], encoding, chartType } = advancedVSeed
 
   if (measures.length > 2) {
-    throw new Error('measures can not be more than 2')
+    throw new Error('measures can not be more than 2 groups in scatter')
   }
   const foldInfoList: FoldInfo[] = []
   const unfoldInfoList: UnfoldInfo[] = []