+++ packages/vseed/src/pipeline/spec/chart/pipes/axes/yBand.ts
@@ -1,10 +1,12 @@
 import type { ICartesianBandAxisSpec, ISpec } from '@visactor/vchart'
 import type { SpecPipe, XBandAxis } from 'src/types'
+import { defaultTitleText } from './title/defaultTitleText'
 
 export const yBand: SpecPipe = (spec, context) => {
   const result = { ...spec } as ISpec
   const { advancedVSeed, vseed } = context
   const { chartType } = vseed
+  const { measures, dimensions, encoding } = advancedVSeed
   const config = advancedVSeed.config?.[chartType as 'bar']?.yAxis as XBandAxis
 
   if (!result.axes) {
@@ -71,7 +73,7 @@ export const yBand: SpecPipe = (spec, context) => {
     },
     title: {
       visible: title?.visible,
-      text: title?.titleText,
+      text: title?.titleText || defaultTitleText(measures, dimensions, encoding.y as string[]),
       style: {
         fill: title?.titleColor,
         fontSize: title?.titleFontSize,