+++ packages/vseed/tests/builder/error/errorDataset.test.ts
@@ -4,11 +4,11 @@ import type { ISpec } from '@visactor/vchart'
 import { VChart } from '@visactor/vchart'
 import * as VTable from '@visactor/vtable'
 
-import vseed from './errorChartType.json'
+import vseed from './errorDataset.json'
 
 VTable.register.chartModule('vchart', VChart)
 
-test('errorChartType', () => {
+test('errorDataset', () => {
   registerAll()
 
   try{
@@ -64,7 +64,7 @@ test('errorChartType', () => {
     }
   } catch (e) {
     console.error(e)
-    expect({expectError: true}).toMatchSnapshot()
+    expect({expectError: true, message: (e as Error).message}).toMatchSnapshot()
     expect(e).toBeInstanceOf(Error)
   }
 })