+++ packages/vseed/tests/encoding/color/column_parallel_multi_measure.test.ts
@@ -64,7 +64,7 @@ test('column_parallel_multi_measure', () => {
     }
   } catch (e) {
     console.error(e)
-    expect({expectError: true}).toMatchSnapshot()
+    expect({expectError: true, message: (e as Error).message}).toMatchSnapshot()
     expect(e).toBeInstanceOf(Error)
   }
 })