+++ packages/vseed/tests/features/dualChartType/type_columnParallel_line.test.ts
@@ -64,7 +64,7 @@ test('type_columnParallel_line', () => {
     }
   } catch (e) {
     console.error(e)
-    expect({expectError: true}).toMatchSnapshot()
+    expect({expectError: true, message: (e as Error).message}).toMatchSnapshot()
     expect(e).toBeInstanceOf(Error)
   }
 })