+++ packages/vseed/tests/features/dualChartType/type_column_scatter.test.ts
@@ -64,7 +64,7 @@ test('type_column_scatter', () => {
     }
   } catch (e) {
     console.error(e)
-    expect({expectError: true}).toMatchSnapshot()
+    expect({expectError: true, message: (e as Error).message}).toMatchSnapshot()
     expect(e).toBeInstanceOf(Error)
   }
 })