+++ packages/vseed/tests/pipeline/chartType/rose/simpleRose.test.ts
@@ -64,7 +64,7 @@ test('simpleRose', () => {
     }
   } catch (e) {
     console.error(e)
-    expect({expectError: true}).toMatchSnapshot()
+    expect({expectError: true, message: (e as Error).message}).toMatchSnapshot()
     expect(e).toBeInstanceOf(Error)
   }
 })